Add modules "Cell" & "InitializeCells"
