+++ src/modules/Cell.ts
@@ -1,0 +1,15 @@
+class Cell {
+  isMineHiddenIn: boolean;
+
+  isOpened: boolean;
+
+  isFlagged: boolean;
+
+  constructor() {
+    this.isMineHiddenIn = false;
+    this.isOpened = false;
+    this.isFlagged = false;
+  }
+}
+
+export default Cell;
